Checking privileges.

+++ src/unidaplan/UpdateExperimentSampleComment.java
@@ -23,7 +23,8 @@ import org.json.JSONObject;
 	    request.setCharacterEncoding("utf-8");
 	    String in = request.getReader().readLine();
 	    String status = "ok";
-
+	    String privilege = "n";
+	    
 	    JSONObject  jsonIn = null;	
 	    int newKeyID = -1;
 	    
@@ -54,43 +55,65 @@ import org.json.JSONObject;
 		
 		try {
 		    dBconn.startDB();	   
-			// get the old string key.
-			pStmt = dBconn.conn.prepareStatement(
-					"SELECT note FROM expp_samples WHERE id=?");
-			pStmt.setInt(1,sampleID);
-			int oldKeyID = dBconn.getSingleIntValue(pStmt);
-			
-			// create a new stringkey
-			pStmt = dBconn.conn.prepareStatement( 			
-					 "INSERT INTO string_key_table VALUES (default,?,NOW(),?) RETURNING id");
-			pStmt.setString(1, newComment);
-			pStmt.setInt(2, userID);
-			newKeyID = dBconn.getSingleIntValue(pStmt);
+		    
+		    // Check privileges
+		    // get corresponding experiment ID
+ 			pStmt = dBconn.conn.prepareStatement(
+ 					"SELECT expp_id FROM expp_samples WHERE id=?");
+ 			pStmt.setInt(1,sampleID);
+ 			int experimentID = dBconn.getSingleIntValue(pStmt);
+ 			pStmt.close();
+		 			
+		    pStmt = dBconn.conn.prepareStatement( 	
+					"SELECT getExperimentRights(vuserid:=?,vexperimentid:=?)");
+			pStmt.setInt(1,userID);
+			pStmt.setInt(2,experimentID);
+			privilege = dBconn.getSingleStringValue(pStmt);
 			pStmt.close();
 			
-			pStmt = dBconn.conn.prepareStatement( 			
-					 "INSERT INTO stringtable VALUES(default,?,?,?,NOW())");
-			pStmt.setInt(1,newKeyID);
-			pStmt.setString(2, "none");
-			pStmt.setString(3, newComment);
-			pStmt.executeUpdate();
-			pStmt.close();
 			
-			pStmt = dBconn.conn.prepareStatement(
-					 "UPDATE expp_samples SET note=?, lastUser=? WHERE id=?");
-			pStmt.setInt(1,newKeyID);
-			pStmt.setInt(2,userID);
-			pStmt.setInt(3,sampleID);
-			pStmt.executeUpdate();
-			pStmt.close();
-			
-			if (oldKeyID>0){
+			if (privilege.equals("w")){
+		    
+				// get the old string key.
 				pStmt = dBconn.conn.prepareStatement(
-						 "DELETE FROM string_key_table WHERE id=?");
-				pStmt.setInt(1,oldKeyID);
+						"SELECT note FROM expp_samples WHERE id=?");
+				pStmt.setInt(1,sampleID);
+				int oldKeyID = dBconn.getSingleIntValue(pStmt);
+				
+				// create a new stringkey
+				pStmt = dBconn.conn.prepareStatement( 			
+						 "INSERT INTO string_key_table VALUES (default,?,NOW(),?) RETURNING id");
+				pStmt.setString(1, newComment);
+				pStmt.setInt(2, userID);
+				newKeyID = dBconn.getSingleIntValue(pStmt);
+				pStmt.close();
+				
+				pStmt = dBconn.conn.prepareStatement( 			
+						 "INSERT INTO stringtable VALUES(default,?,?,?,NOW())");
+				pStmt.setInt(1,newKeyID);
+				pStmt.setString(2, "none");
+				pStmt.setString(3, newComment);
 				pStmt.executeUpdate();
 				pStmt.close();
-			}
+				
+				pStmt = dBconn.conn.prepareStatement(
+						 "UPDATE expp_samples SET note=?, lastUser=? WHERE id=?");
+				pStmt.setInt(1,newKeyID);
+				pStmt.setInt(2,userID);
+				pStmt.setInt(3,sampleID);
+				pStmt.executeUpdate();
+				pStmt.close();
+				
+				if (oldKeyID>0){
+					pStmt = dBconn.conn.prepareStatement(
+							 "DELETE FROM string_key_table WHERE id=?");
+					pStmt.setInt(1,oldKeyID);
+					pStmt.executeUpdate();
+					pStmt.close();
+				}		
+			} else {
+		    	response.setStatus(401);
+		    }
 			
 		} catch (SQLException e) {
 			System.err.println("UpdateExperimentProcessComment: Problems with SQL query");
