+++ src/unidaplan/UpdateSTParamGrp.java
@@ -2,10 +2,12 @@ package unidaplan;
 import java.io.IOException;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
+
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+
 import org.json.JSONException;
 import org.json.JSONObject;
 
@@ -49,34 +51,39 @@ import org.json.JSONObject;
 	    
 		
 		try {
-		    dBconn.startDB();	   
+		    dBconn.startDB();
+		    
+		    if (Unidatoolkit.userHasAdminRights(userID, dBconn)){
 
-			// find the stringkey
-			pStmt=dBconn.conn.prepareStatement(
-					"SELECT stringkey FROM ot_parametergrps WHERE id=?");
-			pStmt.setInt(1,paramGrpID);
-			int stringKey=dBconn.getSingleIntValue(pStmt);
-			pStmt.close();
+
+				// find the stringkey
+				pStmt=dBconn.conn.prepareStatement(
+						"SELECT stringkey FROM ot_parametergrps WHERE id=?");
+				pStmt.setInt(1,paramGrpID);
+				int stringKey=dBconn.getSingleIntValue(pStmt);
+				pStmt.close();
+				
+				// delete old entries in the same language
+				pStmt=dBconn.conn.prepareStatement(
+						"DELETE FROM stringtable WHERE language=? AND string_key=?");
+				pStmt.setString(1,language);
+				pStmt.setInt(2,stringKey);
+				pStmt.executeUpdate();
+				pStmt.close();
+				
 			
-			// delete old entries in the same language
-			pStmt=dBconn.conn.prepareStatement(
-					"DELETE FROM stringtable WHERE language=? AND string_key=?");
-			pStmt.setString(1,language);
-			pStmt.setInt(2,stringKey);
-			pStmt.executeUpdate();
-			pStmt.close();
-			
-		
-			// create database entry for the new name
-			pStmt= dBconn.conn.prepareStatement( 			
-					 "INSERT INTO stringtable VALUES(default,?,?,?,NOW(),?)");
-			pStmt.setInt(1,stringKey);
-			pStmt.setString(2, language);
-			pStmt.setString(3, newName);
-			pStmt.setInt(4,userID);
-			pStmt.executeUpdate();
-			pStmt.close();
-			
+				// create database entry for the new name
+				pStmt= dBconn.conn.prepareStatement( 			
+						 "INSERT INTO stringtable VALUES(default,?,?,?,NOW(),?)");
+				pStmt.setInt(1,stringKey);
+				pStmt.setString(2, language);
+				pStmt.setString(3, newName);
+				pStmt.setInt(4,userID);
+				pStmt.executeUpdate();
+				pStmt.close();
+		    } else {
+		    	response.setStatus(401);
+		    }
 			
 		} catch (SQLException e) {
 			System.err.println("UpdateSTParamGrp: Problems with SQL query");
