+++ src/unidaplan/UpdateSearchName.java
@@ -2,10 +2,12 @@ package unidaplan;
 import java.io.IOException;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
+
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+
 import org.json.JSONException;
 import org.json.JSONObject;
 
@@ -19,6 +21,7 @@ import org.json.JSONObject;
 		int userID=authentificator.GetUserID(request,response);
 	    request.setCharacterEncoding("utf-8");
 	    String in = request.getReader().readLine();
+	    String privilege = "n";
 	    String status = "ok";
 
 	    JSONObject  jsonIn = null;	
@@ -49,34 +52,46 @@ import org.json.JSONObject;
 	    
 		
 		try {
-		    dBconn.startDB();	   
+		    dBconn.startDB();
+		    
+		    // Check privileges
+		    pStmt = dBconn.conn.prepareStatement( 	
+					"SELECT getSearchRights(vuserid:=?,vsearchid:=?)");
+			pStmt.setInt(1,userID);
+			pStmt.setInt(2,searchID);
+			privilege = dBconn.getSingleStringValue(pStmt);
+			pStmt.close();
+						
+			if (privilege.equals("w")){
 
-			// find the stringkey
-			pStmt=dBconn.conn.prepareStatement(
-					"SELECT name FROM searches WHERE id=?");
-			pStmt.setInt(1,searchID);
-			int stringKey=dBconn.getSingleIntValue(pStmt);
-			pStmt.close();
+				// find the stringkey
+				pStmt=dBconn.conn.prepareStatement(
+						"SELECT name FROM searches WHERE id=?");
+				pStmt.setInt(1,searchID);
+				int stringKey=dBconn.getSingleIntValue(pStmt);
+				pStmt.close();
+				
+				// delete old entries in the same language
+				pStmt=dBconn.conn.prepareStatement(
+						"DELETE FROM stringtable WHERE language=? AND string_key=?");
+				pStmt.setString(1,language);
+				pStmt.setInt(2,stringKey);
+				pStmt.executeUpdate();
+				pStmt.close();
+				
 			
-			// delete old entries in the same language
-			pStmt=dBconn.conn.prepareStatement(
-					"DELETE FROM stringtable WHERE language=? AND string_key=?");
-			pStmt.setString(1,language);
-			pStmt.setInt(2,stringKey);
-			pStmt.executeUpdate();
-			pStmt.close();
-			
-		
-			// create database entry for the new name
-			pStmt= dBconn.conn.prepareStatement( 			
-					 "INSERT INTO stringtable VALUES(default,?,?,?,NOW(),?)");
-			pStmt.setInt(1,stringKey);
-			pStmt.setString(2, language);
-			pStmt.setString(3, newName);
-			pStmt.setInt(4,userID);
-			pStmt.executeUpdate();
-			pStmt.close();
-			
+				// create database entry for the new name
+				pStmt= dBconn.conn.prepareStatement( 			
+						 "INSERT INTO stringtable VALUES(default,?,?,?,NOW(),?)");
+				pStmt.setInt(1,stringKey);
+				pStmt.setString(2, language);
+				pStmt.setString(3, newName);
+				pStmt.setInt(4,userID);
+				pStmt.executeUpdate();
+				pStmt.close();
+			} else {
+				response.setStatus(401);
+			}
 			
 		} catch (SQLException e) {
 			System.err.println("UpdateSearchName: Problems with SQL query");
